requirements.txt
absl-py==2.1.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mediapipe==0.10.15
ml_dtypes==0.5.0
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pillow==11.0.0
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.14.1
six==1.16.0
sounddevice==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.6
